FacadeModele::obtenirCentre de masse.
La convertion implicite du centre de masse (qui est des doubles) en int, je l'ai rendu explicite pour pas que ça génère de warnings.

+++ Sources/DLL/Application/FacadeModele.cpp
@@ -883,7 +883,7 @@ int FacadeModele::obtenirCentreMasseX(){
 	int centreMasseX = 0;
 	VisiteurCentreDeMasse* visiteur = new VisiteurCentreDeMasse();
 	FacadeModele::obtenirInstance()->obtenirArbreRenduINF2990()->accepterVisiteur(visiteur);
-	centreMasseX = visiteur->obtenirCentreDeMasse().x;
+	centreMasseX = (int) visiteur->obtenirCentreDeMasse().x;
 	return centreMasseX;
 }
 
@@ -901,7 +901,7 @@ int FacadeModele::obtenirCentreMasseY(){
 	int centreMasseY = 0;
 	VisiteurCentreDeMasse* visiteur = new VisiteurCentreDeMasse();
 	FacadeModele::obtenirInstance()->obtenirArbreRenduINF2990()->accepterVisiteur(visiteur);
-	centreMasseY = visiteur->obtenirCentreDeMasse().y;
+	centreMasseY = (int) visiteur->obtenirCentreDeMasse().y;
 	return centreMasseY;
 }
